+++ subsystems/housing/engines/__init__.py
@@ -1,0 +1,1 @@
+"""Private engines composed by HousingSubsystem."""